history save

+++ src/main/java/service/repository/base/CrudRepository.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface CrudRepository<E> {
 
     void save(E e);
-    E find(int id);
+    E find(Long id);
     List<E> findAll();
-    boolean update(int id);
-    boolean delete(int id);
+    boolean update(Long id);
+    boolean delete(Long id);
 
 }
